libstagefright: start collecting stats after first frame
- AwesomePlayer was collecting statistics starting from first read
  call to first frame, this puts first frame latency into
  statistics consideration which is not correct behavior
- Fixes this problem by waiting until first frame is read and then
  start collecting statistics on FPS
- Also improve precision by using double to store intermediate fps
  values and extending duration to 250000us

Crs-Fixed: 338656

+++ media/libstagefright/include/AwesomePlayer.h
@@ -340,7 +340,7 @@ private:
         int64_t mFirstFrameLatencyStartUs; //first frame latency start
         int64_t mLastFrame;
         int64_t mLastFrameUs;
-        float mFPSSumUs;
+        double mFPSSumUs;
         int64_t mStatisticsFrames;
         bool mVeryFirstFrame;
         int64_t mTotalTime;
